fix jetty security due to some refactoring in 12.0.0-SNAPSHOT
