+++ pom.xml
@@ -48,7 +48,7 @@
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.release>8</maven.compiler.release>
 
-    <jetty12.version>12.0.0.beta0</jetty12.version>
+    <jetty12.version>12.0.0-SNAPSHOT</jetty12.version>
 
   </properties>
